etl-pipeline/requirements.txt
# Database
psycopg2-binary==2.9.5

# Data processing
numpy==1.23.5           # ✅ Install before pandas to avoid binary mismatch
pandas==1.5.3
pyarrow==11.0.0

# Google Cloud
google-cloud-storage==2.8.0
google-cloud-bigquery==3.9.0

# Kubernetes
kubernetes==26.1.0

# Monitoring
prometheus-client==0.16.0

# Utilities
pyyaml==6.0

